requirements.txt
asgiref==3.11.0
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
gunicorn==23.0.0
mysqlclient==2.2.7
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
